Refactor DLL search to be part of __init__ module
